FIX : change DB schema to BIGINT for id fields to match JPA Long mapping

+++ schema.sql
@@ -6,7 +6,7 @@ USE goodjob;
 
 -- 사용자 정보 테이블
 CREATE TABLE users (
-                       id INT PRIMARY KEY AUTO_INCREMENT,                 -- 사용자 고유 ID
+                       id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- 사용자 고유 ID
                        email VARCHAR(255) UNIQUE NOT NULL,                -- 사용자 이메일 (고유값)
                        name VARCHAR(100) NOT NULL,                        -- 사용자 이름
                        region VARCHAR(100),                               -- 사용자의 지역 정보 (예: 경기 수원)
@@ -17,8 +17,8 @@ CREATE TABLE users (
 
 -- OAuth 로그인 정보 테이블 (Google, Kakao 등)
 CREATE TABLE user_oauth (
-                            id INT PRIMARY KEY AUTO_INCREMENT,                 -- OAuth 레코드 고유 ID
-                            user_id INT NOT NULL,                              -- users 테이블의 외래 키
+                            id BIGINT PRIMARY KEY AUTO_INCREMENT,              -- OAuth 레코드 고유 ID
+                            user_id BIGINT NOT NULL,                           -- users 테이블의 외래 키
                             provider VARCHAR(20) NOT NULL,                     -- OAuth 제공자 (예: google, kakao)
                             oauth_id VARCHAR(255) NOT NULL,                    -- OAuth 고유 ID (소셜 계정 ID)
                             access_token TEXT,                                 -- 액세스 토큰
@@ -30,8 +30,8 @@ CREATE TABLE user_oauth (
 
 -- 사용자 CV 테이블 (PDF 업로드 기반)
 CREATE TABLE cv (
-                    id INT PRIMARY KEY AUTO_INCREMENT,                 -- CV 고유 ID
-                    user_id INT NOT NULL,                              -- users 테이블 외래 키
+                    id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- CV 고유 ID
+                    user_id BIGINT NOT NULL,                              -- users 테이블 외래 키
                     file_name VARCHAR(255) NOT NULL,                   -- 업로드한 파일 이름
                     file_url VARCHAR(500) NOT NULL,                    -- S3 등 외부 저장소의 파일 URL
                     raw_text TEXT NOT NULL,                            -- 추출된 전체 CV 텍스트 (OCR 등)
@@ -46,7 +46,7 @@ CREATE TABLE cv (
 
 -- 채용 공고 테이블
 CREATE TABLE jobs (
-                      id INT PRIMARY KEY AUTO_INCREMENT,                 -- 공고 고유 ID
+                      id BIGINT PRIMARY KEY AUTO_INCREMENT,              -- 공고 고유 ID
                       company_name VARCHAR(255) NOT NULL,                -- 회사명
                       job_title VARCHAR(255) NOT NULL,                   -- 직무 제목
                       department VARCHAR(255),                           -- 부서명
@@ -68,9 +68,9 @@ CREATE TABLE jobs (
 
 -- 피드백 테이블 (CV와 채용공고의 1:1 매칭 피드백)
 CREATE TABLE cv_feedback (
-                             id INT PRIMARY KEY AUTO_INCREMENT,                 -- 피드백 고유 ID
-                             cv_id INT NOT NULL,                                -- 어떤 이력서에 대한 피드백인지
-                             job_id INT NOT NULL,                               -- 어떤 공고와 비교한 피드백인지
+                             id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- 피드백 고유 ID
+                             cv_id BIGINT NOT NULL,                                -- 어떤 이력서에 대한 피드백인지
+                             job_id BIGINT NOT NULL,                               -- 어떤 공고와 비교한 피드백인지
                              feedback TEXT NOT NULL,                            -- 부족한 점 등 피드백 내용
                              created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,    -- 피드백 생성 시간
                              FOREIGN KEY (cv_id) REFERENCES cv(id) ON DELETE CASCADE,
@@ -80,9 +80,9 @@ CREATE TABLE cv_feedback (
 
 -- 북마크 테이블
 CREATE TABLE bookmarks (
-                           id INT PRIMARY KEY AUTO_INCREMENT,                 -- 북마크 고유 ID
-                           user_id INT NOT NULL,                              -- 북마크한 사용자
-                           job_id INT NOT NULL,                               -- 북마크한 공고
+                           id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- 북마크 고유 ID
+                           user_id BIGINT NOT NULL,                              -- 북마크한 사용자
+                           job_id BIGINT NOT NULL,                               -- 북마크한 공고
                            saved_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,      -- 북마크 시점
                            FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE,
                            FOREIGN KEY (job_id) REFERENCES jobs(id) ON DELETE CASCADE,
@@ -91,9 +91,9 @@ CREATE TABLE bookmarks (
 
 -- 지원 이력 테이블
 CREATE TABLE applications (
-                              id INT PRIMARY KEY AUTO_INCREMENT,                 -- 지원 이력 고유 ID
-                              user_id INT NOT NULL,                              -- 지원한 사용자
-                              job_id INT NOT NULL,                               -- 지원한 공고
+                              id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- 지원 이력 고유 ID
+                              user_id BIGINT NOT NULL,                              -- 지원한 사용자
+                              job_id BIGINT NOT NULL,                               -- 지원한 공고
                               apply_status ENUM('지원', '서류전형', '코테', '면접', '최종합격', '불합격') DEFAULT '지원', -- 상태
                               applied_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,    -- 최초 지원일
                               note TEXT,                                          -- 사용자가 남긴 메모 또는 후기
@@ -103,8 +103,8 @@ CREATE TABLE applications (
 
 -- 관리자 로그 테이블 (작업 내역 기록용)
 CREATE TABLE admin_logs (
-                            id INT PRIMARY KEY AUTO_INCREMENT,                 -- 로그 고유 ID
-                            admin_id INT NOT NULL,                             -- 작업한 관리자 ID
+                            id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- 로그 고유 ID
+                            admin_id BIGINT NOT NULL,                             -- 작업한 관리자 ID
                             action_type ENUM('USER_DELETE', 'JOB_UPDATE', 'FEEDBACK_REVIEW', 'SYSTEM_LOG', 'JOB_REGISTER'),
     -- 작업 종류
                             details TEXT,                                       -- 작업 상세 설명
@@ -114,8 +114,8 @@ CREATE TABLE admin_logs (
 
 -- 사용자 검색 히스토리 테이블
 CREATE TABLE search_history (
-                                id INT PRIMARY KEY AUTO_INCREMENT,                 -- 검색 기록 ID
-                                user_id INT NOT NULL,                              -- 검색한 사용자
+                                id BIGINT PRIMARY KEY AUTO_INCREMENT,                 -- 검색 기록 ID
+                                user_id BIGINT NOT NULL,                              -- 검색한 사용자
                                 keyword VARCHAR(255) NOT NULL,                     -- 검색어
                                 searched_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,   -- 검색 시점
                                 FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
